don't retrieve team relationship if not using teams
Hi, I just upgraded to Laratrust 5.0.6 and found a small bug. I'm using barryvdh's ide-helper, and when the composer afterUpdate method ran, ide-helper tried to retrieve all the Eloquent relationships, and crashed when it got to the rolesTeams relationship, since I am not using teams and I don't have that model.