Update teams.md
Added a helpful but hidden method, I needed such a method but as it was never mentioned any where in the documentation i thought it didn't exist till i found a commit that specifically mentioned it. If this is not the appropriate section for such a method consider moving to different section but methods such as `$user->roles` and `$user->rolesTeams` really need to be on the documentation. Thank you for all the hard work.